// frontend/src/api/generated/model/protocolViewModel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SmartGrao.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { Crop } from './crop';
import type { ProtocolItemViewModel } from './protocolItemViewModel';
import type { ProtocolStatus } from './protocolStatus';

export interface ProtocolViewModel {
  id: string;
  code: string;
  version: number;
  crop: Crop;
  name: string;
  status: ProtocolStatus;
  items: ProtocolItemViewModel[];
}