// frontend/src/api/generated/model/recordObservationViewModel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SmartGrao.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { GeoCoordinateViewModel } from './geoCoordinateViewModel';
import type { ObservationCountViewModel } from './observationCountViewModel';

export interface RecordObservationViewModel {
  /** @nullable */
  samplingPointId: string | null;
  recordedAt: string;
  location: GeoCoordinateViewModel;
  /** @nullable */
  accuracyMeters: number | null;
  /** @nullable */
  growthStage: string | null;
  /** @nullable */
  notes: string | null;
  counts: ObservationCountViewModel[];
}